+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v0.6.0 (unreleased)
 
-* Update `tokio` dependency to version `0.3.0`
+* Update `tokio` dependency to version `0.3`
+* Update `crossbeam-queue` dependency to version `0.3`
 
 ## v0.5.2